Revert "Merge pull request #1 from Sychara/Vasyan"
This reverts commit a06b649e5f02a5f6dbc6d38903626352832e6e70, reversing
changes made to 460a658ae56252b463f5b5c7b823966331ec6087.

+++ src/Layers/xrRender/dxConsoleRender.cpp
@@ -14,7 +14,7 @@ void dxConsoleRender::OnRender(bool bGame)
 {
     VERIFY(HW.pDevice);
 
-    D3DRECT R = {0, 0, static_cast<signed int>(Device.dwWidth), static_cast<signed int>(Device.dwHeight)};
+    D3DRECT R = {0, 0, static_cast<LONG>(Device.dwWidth), static_cast<LONG>(Device.dwHeight)};
     if (bGame)
         R.y2 /= 2;
